[7532] Avoid warnings at use size_t with printf fromat strings. Author: VladimirMangos
--HG--
branch : trunk

+++ src/game/AchievementMgr.cpp
@@ -1310,7 +1310,7 @@ void AchievementGlobalMgr::LoadAchievementCriteriaList()
     }
 
     sLog.outString("");
-    sLog.outString(">> Loaded %u achievement criteria.",m_AchievementCriteriasByType->size());
+    sLog.outString(">> Loaded %lu achievement criteria.",(unsigned long)m_AchievementCriteriasByType->size());
 }
 
 
@@ -1339,7 +1339,7 @@ void AchievementGlobalMgr::LoadCompletedAchievements()
     delete result;
 
     sLog.outString("");
-    sLog.outString(">> Loaded %u realm completed achievements.",m_allCompletedAchievements.size());
+    sLog.outString(">> Loaded %lu realm completed achievements.",(unsigned long)m_allCompletedAchievements.size());
 }
 
 void AchievementGlobalMgr::LoadRewards()
@@ -1449,7 +1449,7 @@ void AchievementGlobalMgr::LoadRewards()
     delete result;
 
     sLog.outString("");
-    sLog.outString( ">> Loaded %u achievement reward locale strings", m_achievementRewardLocales.size() );
+    sLog.outString( ">> Loaded %lu achievement reward locale strings", (unsigned long)m_achievementRewardLocales.size() );
 }
 
 void AchievementGlobalMgr::LoadRewardLocales()
@@ -1517,6 +1517,7 @@ void AchievementGlobalMgr::LoadRewardLocales()
 
     delete result;
 
+
     sLog.outString("");
-    sLog.outString( ">> Loaded %u achievement reward locale strings", m_achievementRewardLocales.size() );
+    sLog.outString( ">> Loaded %lu achievement reward locale strings", (unsigned long)m_achievementRewardLocales.size() );
 }
